Clear console.log and import format
