+++ server/src/answer/answer.controller.ts
@@ -12,12 +12,12 @@ import {
 } from "@nestjs/common";
 import { Response } from "express";
 import { AnswerService } from "./answer.service";
-import { JwtAuthGuard } from "src/auth/jwt-auth.guard";
+import { JwtAuthGuard } from "../auth/jwt-auth.guard";
 import { FilesInterceptor } from "@nestjs/platform-express";
-import { FileService } from "src/file/file.service";
+import { FileService } from "../file/file.service";
 import { CreateAnswerDto } from "./dto/createAnswer.dto";
-import { UserService } from "src/user/user.service";
-import { QuestionService } from "src/question/question.service";
+import { UserService } from "../user/user.service";
+import { QuestionService } from "../question/question.service";
 import { InteractAnswerQueryDto } from "./dto/interactQuery.dto";
 
 @Controller("answers")
@@ -75,8 +75,6 @@ export class AnswerController {
 			};
 			const answer = await this.answerService.createAnswer(query);
 
-			// Background
-
 			this.userService.answeredCompute(req.user?._id);
 			this.questionService.IncreaseAnsweredAndParticipant(
 				questionDetail?._id,
@@ -89,8 +87,6 @@ export class AnswerController {
 			if (files.length > 0) {
 				await this.fileService.removeFiles(uploadedFiles);
 			}
-			console.log("error from answer controller create answer function.");
-			console.log(err);
 			throw new InternalServerErrorException("Something went wrong.");
 		}
 	}
@@ -112,7 +108,6 @@ export class AnswerController {
 				reputation: isLike ? 1 : -1,
 			};
 
-			// Background
 			this.questionService.increaseParticipant(
 				answer.answeredIn.toString(),
 			);
@@ -122,8 +117,6 @@ export class AnswerController {
 				success: true,
 			});
 		} catch (err) {
-			console.log("error from answer controller like answer function/");
-			console.log(err);
 			throw new InternalServerErrorException("Something went wrong.");
 		}
 	}
