+++ server/src/answer/answer.module.ts
@@ -3,9 +3,9 @@ import { AnswerController } from "./answer.controller";
 import { AnswerService } from "./answer.service";
 import { MongooseModule } from "@nestjs/mongoose";
 import { Answer, AnswerSchema } from "./schema/answer.schema";
-import { FileModule } from "src/file/file.module";
-import { UserModule } from "src/user/user.module";
-import { QuestionModule } from "src/question/question.module";
+import { FileModule } from "../file/file.module";
+import { UserModule } from "../user/user.module";
+import { QuestionModule } from "../question/question.module";
 
 @Module({
 	imports: [
