+++ server/src/auth/local.strategy.ts
@@ -2,7 +2,7 @@ import { Injectable, UnauthorizedException } from "@nestjs/common";
 import { PassportStrategy } from "@nestjs/passport";
 import { Strategy } from "passport-local";
 import { AuthService } from "./auth.service";
-import { UserDetail } from "src/user/interfaces/user.interface";
+import { UserDetail } from "../user/interfaces/user.interface";
 
 @Injectable()
 export class LocalStrategy extends PassportStrategy(Strategy, "login") {
@@ -10,7 +10,10 @@ export class LocalStrategy extends PassportStrategy(Strategy, "login") {
 		super();
 	}
 
-	async validate(username: string, password: string): Promise<UserDetail | null> {
+	async validate(
+		username: string,
+		password: string,
+	): Promise<UserDetail | null> {
 		const user = await this.authService.validateUser(username, password);
 		if (!user) {
 			throw new UnauthorizedException("Your credentials is incorrect");
