+++ server/src/reply/reply.controller.ts
@@ -9,19 +9,17 @@ import {
 } from "@nestjs/common";
 import { Response } from "express";
 import { ReplyService } from "./reply.service";
-import { QuestionService } from "src/question/question.service";
-import { UserService } from "src/user/user.service";
-import { AnswerService } from "src/answer/answer.service";
-import { JwtAuthGuard } from "src/auth/jwt-auth.guard";
+import { QuestionService } from "../question/question.service";
+import { AnswerService } from "../answer/answer.service";
+import { JwtAuthGuard } from "../auth/jwt-auth.guard";
 import { CreateReplyDto } from "./dto/createReply.dto";
-import { PushReplyToAnswer } from "src/dto/pushReplyQuery.dto";
+import { PushReplyToAnswer } from "../answer/dto/pushReplyQuery.dto";
 
 @Controller("replies")
 export class ReplyController {
 	constructor(
 		private replyService: ReplyService,
 		private questionService: QuestionService,
-		private userService: UserService,
 		private answerService: AnswerService,
 	) {}
 
@@ -43,7 +41,6 @@ export class ReplyController {
 				replyId: reply._id.toString(),
 			};
 
-			// Background
 			this.questionService.increaseParticipant(req.body.questionId);
 			this.answerService.pushReplyToAnswer(query);
 
@@ -51,8 +48,6 @@ export class ReplyController {
 				reply: reply,
 			});
 		} catch (err) {
-			console.log("error from reply controller create reply function.");
-			console.log(err);
 			throw new InternalServerErrorException("Something went wrong.");
 		}
 	}
