+++ server/src/reply/reply.module.ts
@@ -3,9 +3,8 @@ import { ReplyController } from "./reply.controller";
 import { ReplyService } from "./reply.service";
 import { MongooseModule } from "@nestjs/mongoose";
 import { Reply, ReplySchema } from "./schema/reply.schema";
-import { AnswerModule } from "src/answer/answer.module";
-import { QuestionModule } from "src/question/question.module";
-import { UserModule } from "src/user/user.module";
+import { AnswerModule } from "../answer/answer.module";
+import { QuestionModule } from "../question/question.module";
 
 @Module({
 	imports: [
@@ -17,7 +16,6 @@ import { UserModule } from "src/user/user.module";
 		]),
 		forwardRef(() => AnswerModule),
 		forwardRef(() => QuestionModule),
-		forwardRef(() => UserModule),
 	],
 	controllers: [ReplyController],
 	providers: [ReplyService],
