+++ server/src/tag/schema/tag.schema.ts
@@ -1,7 +1,7 @@
 import { Prop, Schema, SchemaFactory } from "@nestjs/mongoose";
 import * as mongoose from "mongoose";
 import { Question } from "../../question/schema/question.schema";
-import { User } from "src/user/schema/user.schema";
+import { User } from "../../user/schema/user.schema";
 
 export type TagDocument = mongoose.HydratedDocument<Tag | null>;
