+++ server/src/tag/tag.controller.ts
@@ -13,16 +13,14 @@ import {
 } from "@nestjs/common";
 import { TagService } from "./tag.service";
 import { Response, Request } from "express";
-import { QuestionService } from "src/question/question.service";
-import { PreviewQuestionDto } from "src/question/dto/previewQuestion.dto";
-import { UserService } from "src/user/user.service";
-import {
-	previewQuestionFormat,
-	previewTagFormat,
-} from "src/utils/formatter.utils";
+import { QuestionService } from "../question/question.service";
+import { PreviewQuestionDto } from "../question/dto/previewQuestion.dto";
+import { UserService } from "../user/user.service";
+import { previewQuestionFormat } from "../question/utils/formatter.util";
 import { PreviewTagDto } from "./dto/previewTag.dto";
-import { JwtAuthGuard } from "src/auth/jwt-auth.guard";
-import { shuffle } from "../utils/util.util";
+import { previewTagFormat } from "./utils/formatter.util";
+import { JwtAuthGuard } from "../auth/jwt-auth.guard";
+import { shuffle } from "./utils/shuffle.util";
 
 @Controller("tags")
 export class TagController {
@@ -45,8 +43,6 @@ export class TagController {
 				tags,
 			});
 		} catch (err) {
-			console.log(`error from tag controller get all tag`);
-			console.log(err);
 			throw new InternalServerErrorException("Something went wrong");
 		}
 	}
@@ -108,10 +104,6 @@ export class TagController {
 				questions: responses,
 			});
 		} catch (err) {
-			console.log(
-				"error from tag controller filter interested tags function.",
-			);
-			console.log(err);
 			throw new InternalServerErrorException("Something went wrong.");
 		}
 	}
@@ -152,8 +144,6 @@ export class TagController {
 				tag: response,
 			});
 		} catch (err) {
-			console.log(`error from tag controller get tag by tag id function`);
-			console.log(err);
 			throw new InternalServerErrorException("Something went wrong.");
 		}
 	}
