+++ server/src/tag/utils/formatter.util.ts
@@ -1,0 +1,14 @@
+import { PreviewTagDto } from "../dto/previewTag.dto";
+
+export function previewTagFormat(
+	tag: any,
+	questions: any,
+): PreviewTagDto | null {
+	return {
+		_id: tag?._id,
+		name: tag?.name,
+		totalQuestion: tag?.questions.length,
+		questions: questions,
+		createdAt: tag?.createdAt,
+	};
+}
