+++ server/src/user/Dto/user.dto.ts
@@ -1,9 +1,4 @@
-import {
-	IsEmail,
-	IsNotEmpty,
-	IsStrongPassword,
-	MinLength,
-} from "class-validator";
+import { IsNotEmpty, MinLength } from "class-validator";
 import { TagInterface } from "../../tag/interfaces/tag.interface";
 
 export class CreateUserDto {
@@ -15,11 +10,9 @@ export class CreateUserDto {
 	username: string;
 
 	@IsNotEmpty({ message: "Please provide your password" })
-	// @IsStrongPassword()
 	@MinLength(6)
 	password: string;
 
-	@IsEmail()
 	email: string;
 
 	tags: TagInterface[];
@@ -33,7 +26,6 @@ export class EditUserDto {
 	@IsNotEmpty({ message: "Please provide your name" })
 	name: string;
 
-	@IsEmail()
 	email: string;
 
 	bio: string;
@@ -47,6 +39,5 @@ export class EditUserDto {
 
 export class ChangePasswordDto {
 	@IsNotEmpty({ message: "Please provide your password" })
-	@IsStrongPassword()
 	password: string;
 }
