+++ server/src/user/Schema/user.schema.ts
@@ -1,6 +1,6 @@
 import { Prop, Schema, SchemaFactory } from "@nestjs/mongoose";
 import * as mongoose from "mongoose";
-import { Tag, TagSchema } from "src/tag/schema/tag.schema";
+import { Tag } from "../../tag/schema/tag.schema";
 
 export type UserDocument = mongoose.HydratedDocument<User | null>;
