+++ server/src/user/utils/cookie.util.ts
@@ -3,7 +3,6 @@ import { Response } from "express";
 export function attachCookie(res: Response, token: string) {
 	const oneDay = 1000 * 60 * 60 * 24; // millisecond
 	res.cookie("accessToken", token, {
-		// httpOnly: true,
 		expires: new Date(Date.now() + oneDay),
 	});
 }
